+++ packages/@tinacms/form-builder/src/index.ts
@@ -20,4 +20,5 @@ export * from './field-plugin'
 export * from './form-builder'
 export * from './fields-builder'
 export * from './Form'
+export * from './FormMaker'
 export { FieldRenderProps } from 'react-final-form'
